Add PIXe 1084 driver to ./prober/instruments
Adds the PIXe driver to the example instruments, as well as specifying it as a file to be loaded in the init.py file for instruments. Played around with a new idea here: making the drivers for the instrument interfaceable from the command line. You can tell the command line the function you want to run, and the file will do so. May possibly make testing and troubleshooting a bit simpler.