Add PIXe4140 and rename 1084 to 4322
Adds the driver for the SMU file and renames the NI PIXe-1084 file to 4322 to match the actual device the file drives. The SMU drives has capability to just:
1.) set voltage
2.) set current
3.) set voltage get voltage
4.) set votlage get current
5.) set current get voltage
6.) set current get current
The init files were changes as well, and the requirements.txt was also altered to add the requirements for the two files in question.